commit changes

+++ projectgmaps/Models/FarmerLocation.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.ComponentModel.DataAnnotations;
 using System.Linq;
 using System.Threading.Tasks;
 
@@ -7,11 +8,12 @@ namespace projectgmaps.Models
 {
     public class FarmerLocation
     {
+       // [Key]
 
         public int Id { get; set; }
         public string Name { get; set; }
-        public float Latitude { get; set; }
-        public float Longitude { get; set; }
+        public decimal Latitude { get; set; }
+        public decimal Longitude { get; set; }
         public string Address { get; set; }
         public bool IsFlagged { get; set; }
